+++ go_services/historialFotos/main.go
@@ -4,11 +4,43 @@ import (
 	"fmt"
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/dynamodb"
+	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbattribute"
 )
 
+// Item Create struct to hold info about new item
+type Item struct {
+	id      int
+	nombre  string
+	cancion string
+}
+
 func handler(request events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error) {
+	sess := session.Must(session.NewSessionWithOptions(session.Options{
+		SharedConfigState: session.SharedConfigEnable,
+	}))
+	svc := dynamodb.New(sess)
+
+	input := &dynamodb.GetItemInput{
+		Key: map[string]*dynamodb.AttributeValue{
+			"id": {
+				N: aws.String("666"),
+			},
+		},
+		TableName: aws.String("HistorialFotosEmociones"),
+	}
+
+	result, err2 := svc.GetItem(input)
+	if err2 != nil {
+		fmt.Println(err2.Error)
+		fmt.Println("este es un error")
+	}
+	item := Item{}
+	dynamodbattribute.UnmarshalMap(result.Item, &item)
 	return events.APIGatewayProxyResponse{
-		Body:       fmt.Sprintf("Prueba desde AWS Lambda con GO"),
+		Body:       fmt.Sprintf(item.cancion),
 		StatusCode: 200,
 	}, nil
 }
